# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nGX API
  version: 1.0.0
  description: |
    Email infrastructure purpose-built for AI agents.

    nGX lets you provision inboxes, send and receive email, manage
    drafts with human-in-the-loop review, subscribe to real-time events via
    webhooks or WebSockets, and search across your entire message history.

    ## Authentication
    All API requests (except `/health`) require a Bearer API key in the
    `Authorization` header:
    ```
    Authorization: Bearer am_live_...
    ```
    Keys are scoped — see the `scopes` field on the key endpoints.

    ## Organization And Keys
    Organizations are created during account signup and are discovered at runtime
    via the authenticated API key. `GET /org` returns the current organization
    for that key.

    `POST /keys` creates additional API keys for the current organization.
    It does not create a new organization or perform account registration.
  contact:
    name: nGX Support
    url: https://yourdomain.com
servers:
  - url: https://api.yourdomain.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /webhooks:
    get:
      tags: [Webhooks]
      operationId: listWebhooks
      summary: List webhooks
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Webhooks]
      operationId: createWebhook
      summary: Create a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /webhooks/{webhookId}:
    parameters:
      - $ref: "#/components/parameters/webhookId"
    get:
      tags: [Webhooks]
      operationId: getWebhook
      summary: Get a webhook
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Webhooks]
      operationId: updateWebhook
      summary: Update a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebhookRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Webhooks]
      operationId: deleteWebhook
      summary: Delete a webhook
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "API Key (am_live_...)"
      description: |
        All API requests require a Bearer API key.
        Keys can be scoped to specific capabilities and pods.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
...
    webhookId:
      name: webhookId
      in: path
      required: true
      description: Webhook ID.
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    pageToken:
      name: page_token
      in: query
      description: Opaque page token from a previous response.
      schema:
        type: string

  # -------------------------------------------------------------------------
  # Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad Request — invalid parameters or request body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "validation failed"
            details:
              field: "url"
              reason: "must be a valid https URL"
    Unauthorized:
      description: Unauthorized — missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "missing or invalid authorization header"
...
    NotFound:
      description: Not Found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
    Conflict:
      description: Conflict — resource already exists or state transition is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "email address already registered"
...
    InternalError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "not found"
        details:
          type: object
          description: Optional structured error details (validation errors, etc.).
          additionalProperties: true
      example:
        error: "not found"
        details:
          resource: "thread"
          id: "9b6ba4f5-9f7d-4f56-b0ec-f3a0d5f3a201"
    OperationStatus:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        resource:
          type: string
        id:
          type: string
          format: uuid
      example:
        success: true
        message: "resource deleted"
        resource: "inbox"
        id: "3f2d9d77-89b9-4b8f-9f47-d0b78664b6cb"
...
    WebhookAuthHeader:
      type: object
      description: |
        Caller-supplied HTTP header that nGX injects on every outbound
        delivery call to the webhook URL. Lets the receiving server authenticate
        nGX's requests against its own auth system (e.g. a Bearer token or
        API key). The `value` field is write-only — it is accepted on create/update
        but **never returned** in any response. Only `name` is returned so callers
        can confirm a header is configured without exposing the secret.
      required: [name]
      properties:
        name:
          type: string
          description: HTTP header name, e.g. `Authorization` or `X-Api-Key`.
          example: Authorization
        value:
          type: string
          description: Header value (write-only). Not returned in responses.
          example: "Bearer tok_live_abc123"
          writeOnly: true
    Webhook:
      type: object
      required: [webhook_id, url, event_types, enabled, created_at, updated_at]
      properties:
        webhook_id:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        event_types:
          type: array
          items:
            type: string
        pod_ids:
          type: array
          items:
            type: string
        inbox_ids:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        auth_header:
          allOf:
            - $ref: "#/components/schemas/WebhookAuthHeader"
          nullable: true
          description: |
            Present (with only `name`) when a caller-supplied auth header is
            configured. Absent when none is set.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        webhook_id: "wh_123"
        url: "https://example.com/webhooks/nGX"
        description: "Production agent automation"
        event_types:
          - "message.received"
          - "draft.created"
        pod_ids:
          - "pod_123"
        inbox_ids: []
        enabled: true
        created_at: "2026-03-27T18:15:00Z"
        updated_at: "2026-03-27T18:20:00Z"
    WebhookList:
      type: object
      required: [count, webhooks]
      properties:
        count:
          type: integer
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        webhooks:
          - webhook_id: "wh_123"
            url: "https://example.com/webhooks/nGX"
            description: "Production agent automation"
            event_types:
              - "message.received"
            pod_ids:
              - "pod_123"
            inbox_ids: []
            enabled: true
            created_at: "2026-03-27T18:15:00Z"
            updated_at: "2026-03-27T18:20:00Z"
        limit: 20
        next_page_token: null
    CreateWebhookRequest:
      type: object
      required: [url, event_types]
      properties:
        url:
          type: string
          format: uri
        description:
          type: string
        event_types:
          type: array
          minItems: 1
          items:
            type: string
        pod_ids:
          type: array
          items:
            type: string
        inbox_ids:
          type: array
          items:
            type: string
        auth_header:
          $ref: "#/components/schemas/WebhookAuthHeader"
          description: |
            Optional. Provide both `name` and `value`. nGX encrypts the
            value at rest and injects the header on every delivery call.
            The `secret` returned on creation is separate — it is used for
            HMAC payload signing so the receiver can verify the call came from
            nGX.
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        description:
          type: string
        event_types:
          type: array
          minItems: 1
          items:
            type: string
        pod_ids:
          type: array
          items:
            type: string
        inbox_ids:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        auth_header:
          $ref: "#/components/schemas/WebhookAuthHeader"
          nullable: true
          description: |
            Omit to leave the existing auth header unchanged.
            Set to `null` to remove the auth header entirely.
            Provide `name` + `value` to replace it.
...